:truck: chore: refactored utils
