+++ app/src/main/java/com/shub39/grit/app/App.kt
@@ -45,6 +45,7 @@ import androidx.navigation3.ui.NavDisplay
 import com.shub39.grit.app.AppSections.Companion.toIconRes
 import com.shub39.grit.app.AppSections.Companion.toStringRes
 import com.shub39.grit.billing.PaywallPage
+import com.shub39.grit.core.LocalWindowSizeClass
 import com.shub39.grit.core.domain.MainAppState
 import com.shub39.grit.core.domain.Sections
 import com.shub39.grit.core.habits.presentation.ui.HabitsGraph
@@ -53,7 +54,6 @@ import com.shub39.grit.core.navigation.verticalTransitionMetadata
 import com.shub39.grit.core.presentation.ChangelogSheet
 import com.shub39.grit.core.presentation.settings.ui.SettingsGraph
 import com.shub39.grit.core.tasks.presentation.ui.TasksPage
-import com.shub39.grit.core.utils.LocalWindowSizeClass
 import com.shub39.grit.viewmodels.HabitViewModel
 import com.shub39.grit.viewmodels.SettingsViewModel
 import com.shub39.grit.viewmodels.TasksViewModel
