+++ app/src/main/java/com/shub39/grit/app/MainActivity.kt
@@ -32,11 +32,11 @@ import androidx.core.content.ContextCompat
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.shub39.grit.core.LocalWindowSizeClass
 import com.shub39.grit.core.data.GritNotificationManager.Companion.createNotificationChannel
 import com.shub39.grit.core.data.Utils
 import com.shub39.grit.core.presentation.component.InitialLoading
 import com.shub39.grit.core.theme.GritTheme
-import com.shub39.grit.core.utils.LocalWindowSizeClass
 import com.shub39.grit.viewmodels.MainViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
