+++ app/src/main/java/com/shub39/grit/core/data/NotificationAlarmScheduler.kt
@@ -25,8 +25,8 @@ import android.util.Log
 import com.shub39.grit.core.domain.AlarmScheduler
 import com.shub39.grit.core.domain.IntentActions
 import com.shub39.grit.core.habits.domain.Habit
+import com.shub39.grit.core.now
 import com.shub39.grit.core.tasks.domain.Task
-import com.shub39.grit.core.utils.now
 import kotlin.time.ExperimentalTime
 import kotlinx.datetime.DateTimeUnit
 import kotlinx.datetime.LocalDateTime
