+++ app/src/main/java/com/shub39/grit/core/presentation/settings/ui/SettingsGraph.kt
@@ -29,6 +29,7 @@ import androidx.navigation3.runtime.NavKey
 import androidx.navigation3.runtime.entryProvider
 import androidx.navigation3.runtime.rememberNavBackStack
 import androidx.navigation3.ui.NavDisplay
+import com.shub39.grit.core.GritPreviewWrapper
 import com.shub39.grit.core.navigation.horizontalTransitionMetadata
 import com.shub39.grit.core.presentation.settings.SettingsAction
 import com.shub39.grit.core.presentation.settings.SettingsState
@@ -37,7 +38,6 @@ import com.shub39.grit.core.presentation.settings.ui.section.Changelog
 import com.shub39.grit.core.presentation.settings.ui.section.LookAndFeelPage
 import com.shub39.grit.core.presentation.settings.ui.section.RootPage
 import com.shub39.grit.core.shared_ui.PageFill
-import com.shub39.grit.core.utils.GritPreviewWrapper
 import kotlinx.serialization.Serializable
 
 @Serializable data object Root : NavKey
