+++ app/src/main/java/com/shub39/grit/habits/data/repository/util.kt
@@ -19,7 +19,7 @@ package com.shub39.grit.habits.data.repository
 import com.shub39.grit.core.habits.domain.HabitStatus
 import com.shub39.grit.core.habits.domain.WeekDayFrequencyData
 import com.shub39.grit.core.habits.domain.WeeklyComparisonData
-import com.shub39.grit.core.utils.now
+import com.shub39.grit.core.now
 import kotlin.time.ExperimentalTime
 import kotlinx.datetime.DatePeriod
 import kotlinx.datetime.DateTimeUnit
