+++ app/src/main/java/com/shub39/grit/widgets/habit_overview_widget/HabitOverviewWidget.kt
@@ -63,7 +63,7 @@ import com.shub39.grit.app.MainActivity
 import com.shub39.grit.core.habits.domain.Habit
 import com.shub39.grit.core.habits.domain.HabitRepo
 import com.shub39.grit.core.habits.domain.HabitStatus
-import com.shub39.grit.core.utils.now
+import com.shub39.grit.core.now
 import com.shub39.grit.widgets.WidgetSize
 import kotlin.time.ExperimentalTime
 import kotlinx.coroutines.launch
