+++ app/src/main/java/com/shub39/grit/widgets/habit_streak_widget/HabitStreakWidget.kt
@@ -67,7 +67,7 @@ import com.shub39.grit.app.MainActivity
 import com.shub39.grit.core.habits.domain.Habit
 import com.shub39.grit.core.habits.domain.HabitRepo
 import com.shub39.grit.core.habits.domain.HabitWithAnalytics
-import com.shub39.grit.core.utils.now
+import com.shub39.grit.core.now
 import com.shub39.grit.widgets.WidgetSize
 import kotlinx.coroutines.launch
 import kotlinx.datetime.LocalDateTime
