+++ shared/core/src/commonMain/kotlin/com/shub39/grit/core/GritPreviewWrapper.kt
@@ -14,6 +14,16 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.shub39.grit.core.utils
+package com.shub39.grit.core
 
-actual fun blurPossible(): Boolean = true
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.PreviewWrapperProvider
+import com.shub39.grit.core.theme.GritTheme
+import com.shub39.grit.core.theme.Theme
+
+class GritPreviewWrapper : PreviewWrapperProvider {
+    @Composable
+    override fun Wrap(content: @Composable (() -> Unit)) {
+        GritTheme(theme = Theme()) { content() }
+    }
+}
