+++ shared/core/src/commonMain/kotlin/com/shub39/grit/core/blurPossible.kt
@@ -14,21 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.shub39.grit.core.utils
+package com.shub39.grit.core
 
-import kotlin.time.Clock
-import kotlinx.datetime.LocalDate
-import kotlinx.datetime.LocalDateTime
-import kotlinx.datetime.LocalTime
-import kotlinx.datetime.TimeZone
-import kotlinx.datetime.toLocalDateTime
-import kotlinx.datetime.todayIn
-
-// Frequently used time utils
-
-fun LocalDateTime.Companion.now(): LocalDateTime =
-    Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault())
-
-fun LocalDate.Companion.now(): LocalDate = Clock.System.todayIn(TimeZone.currentSystemDefault())
-
-fun LocalTime.Companion.now(): LocalTime = LocalDateTime.now().time
+/** is Blur possible on this platform? */
+expect fun blurPossible(): Boolean
