+++ shared/core/src/commonMain/kotlin/com/shub39/grit/core/habits/presentation/ui/component/HabitStreakCard.kt
@@ -52,8 +52,8 @@ import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewWrapper
 import androidx.compose.ui.unit.dp
+import com.shub39.grit.core.GritPreviewWrapper
 import com.shub39.grit.core.theme.flexFontRounded
-import com.shub39.grit.core.utils.GritPreviewWrapper
 import grit.shared.core.generated.resources.Res
 import grit.shared.core.generated.resources.best_streak
 import grit.shared.core.generated.resources.heat
