+++ shared/core/src/commonMain/kotlin/com/shub39/grit/core/habits/presentation/ui/sections/HabitsList.kt
@@ -30,18 +30,18 @@ import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import com.shub39.grit.core.LocalWindowSizeClass
 import com.shub39.grit.core.habits.domain.Habit
 import com.shub39.grit.core.habits.presentation.HabitState
 import com.shub39.grit.core.habits.presentation.HabitsAction
 import com.shub39.grit.core.habits.presentation.ui.component.HabitCard
 import com.shub39.grit.core.habits.presentation.ui.component.HabitUpsertSheet
+import com.shub39.grit.core.now
 import com.shub39.grit.core.shared_ui.Empty
 import com.shub39.grit.core.shared_ui.detachedItemShape
 import com.shub39.grit.core.shared_ui.endItemShape
 import com.shub39.grit.core.shared_ui.leadingItemShape
 import com.shub39.grit.core.shared_ui.middleItemShape
-import com.shub39.grit.core.utils.LocalWindowSizeClass
-import com.shub39.grit.core.utils.now
 import grit.shared.core.generated.resources.Res
 import grit.shared.core.generated.resources.drag_indicator
 import kotlinx.datetime.DayOfWeek
