+++ shared/core/src/commonMain/kotlin/com/shub39/grit/core/tasks/presentation/ui/component/TaskCard.kt
@@ -43,7 +43,7 @@ import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.shub39.grit.core.tasks.domain.Task
-import com.shub39.grit.core.utils.toFormattedString
+import com.shub39.grit.core.toFormattedString
 import grit.shared.core.generated.resources.Res
 import grit.shared.core.generated.resources.alarm
 import org.jetbrains.compose.resources.vectorResource
