+++ shared/core/src/wasmJsMain/kotlin/com/shub39/grit/core/blurPossible.wasmJs.kt
@@ -14,13 +14,6 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.shub39.grit.core.utils
+package com.shub39.grit.core
 
-import androidx.compose.material3.windowsizeclass.WindowSizeClass
-import androidx.compose.runtime.ProvidableCompositionLocal
-import androidx.compose.runtime.staticCompositionLocalOf
-
-/** CompositionLocal for window size class to make adaptive screens */
-val LocalWindowSizeClass: ProvidableCompositionLocal<WindowSizeClass> = staticCompositionLocalOf {
-    error("No window size class provided")
-}
+actual fun blurPossible(): Boolean = true
